+++ CHANGELOG.md
@@ -57,8 +57,11 @@
 * `TelegramotAPI`:
     * Fix: for sending requests caption and text lengths limits were updated
     * New variant of `row` was added
+    * `makeLinkToMessage` extensions has been deprecated (replaced into `TelegramBotAPI-extensions-utils`)
 * `TelegramBotAPI-extensions-utils`:
     * `safely` function was introduced. It is in `PreviewFeature` state currently
+    * `makeLinkToMessage` extensions has been added
+    * `makeLinkToAddStickerSet` function and its variations were added
 
 ### 0.27.4